create enrollment folder and change file names to accommodate
